Release/0.0.4

+++ network/group-chat/src/main/java/com/seugi/network/groupchat/request/GroupChatCreateRequest.kt
@@ -3,6 +3,6 @@ package com.seugi.network.personalchat.request
 data class GroupChatCreateRequest(
     val workspaceId: String,
     val roomName: String,
-    val joinUsers: List<Int>,
+    val joinUsers: List<Long>,
     val chatRoomImg: String,
 )
